requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
sqlalchemy==2.0.36
aiosqlite==0.20.0
pandas==2.2.3
openpyxl==3.1.5
google-genai==1.1.0
python-dotenv==1.0.1
